requirements.txt
requests
pillow
flask
gunicorn